requires #442 and #453 + some rebase and history updated after merging these two.

Reduce allocations in http parser.

benchmarks:

    original:
    BenchmarkHTTPSimpleResponse-4   	  100000	     15197 ns/op	    3552 B/op	     140 allocs/op
    BenchmarkHTTPSimpleRequest-4    	  100000	     17538 ns/op	    4320 B/op	     154 allocs/op
    BenchmarkHTTPSplitResponse-4    	  100000	     15665 ns/op	    3840 B/op	     141 allocs/op
    BenchmarkHttpSimpleTransaction-4	   30000	     44417 ns/op	   11039 B/op	     348 allocs/op

    optimized
    BenchmarkHTTPSimpleResponse-4   	  500000	      2837 ns/op	     560 B/op	       2 allocs/op
    BenchmarkHTTPSimpleRequest-4    	  500000	      3217 ns/op	     640 B/op	       3 allocs/op
    BenchmarkHTTPSplitResponse-4    	 1000000	      3454 ns/op	     848 B/op	       3 allocs/op
    BenchmarkHttpSimpleTransaction-4	  100000	     17065 ns/op	    4398 B/op	      58 allocs/op