@urso 2015-12-08
benchmark results with guarded debug statements:

    BenchmarkHTTPSimpleResponse-4   	 1000000	      2641 ns/op	     560 B/op	       2 allocs/op
    BenchmarkHTTPSimpleRequest-4    	  500000	      3113 ns/op	     640 B/op	       3 allocs/op
    BenchmarkHTTPSplitResponse-4    	  500000	      3223 ns/op	     848 B/op	       3 allocs/op
    BenchmarkHttpSimpleTransaction-4	  100000	     15572 ns/op	    4045 B/op	      50 allocs/op