@andrewkroh 2015-12-08
LGTM. Just needs #442 to merge and then a rebase.